remove member from adyenMerchantPicker
