+++ payment-service/src/main/java/com/hedvig/paymentservice/services/adyen/util/AdyenMerchantPicker.kt
@@ -1,7 +1,7 @@
 package com.hedvig.paymentservice.services.adyen.util
 
 import com.hedvig.paymentservice.configuration.MerchantAccounts
-import com.hedvig.paymentservice.query.member.entities.MemberRepository
+import com.hedvig.paymentservice.query.adyenAccount.AdyenAccountRepository
 import com.hedvig.paymentservice.serviceIntergration.memberService.MemberService
 import com.hedvig.paymentservice.serviceIntergration.productPricing.ProductPricingService
 import com.hedvig.paymentservice.serviceIntergration.underwriterClient.UnderwriterService
@@ -18,7 +18,7 @@ class AdyenMerchantPicker(
   val memberService: MemberService,
   val underwriterService: UnderwriterService,
   val productPricingService: ProductPricingService,
-  val memberRepository: MemberRepository,
+  val adyenAccountRepository: AdyenAccountRepository,
   val merchantAccounts: MerchantAccounts
 ) {
   @Throws(NoMerchantAccountForMarket::class)
@@ -37,13 +37,13 @@ class AdyenMerchantPicker(
   }
 
   private fun getMerchantFromMember(memberId: String): Market? {
-    val memberMaybe = memberRepository.findById(memberId)
+    val accountMaybe = adyenAccountRepository.findById(memberId)
 
-    if (!memberMaybe.isPresent) return null
+    if (!accountMaybe.isPresent) return null
 
-    val member = memberMaybe.get()
+    val account = accountMaybe.get()
 
-    return member.adyenMerchantAccount?.let { adyenMerchantAccount ->
+    return account.merchantAccount?.let { adyenMerchantAccount ->
       Market.valueOf(
         merchantAccounts.merchantAccounts!!
           .filterValues { it == adyenMerchantAccount }.keys.first()
