+++ payment-service/src/test/java/com/hedvig/paymentservice/services/adyen/util/AdyenMerchantPickerTest.kt
@@ -1,8 +1,8 @@
 package com.hedvig.paymentservice.services.adyen.util
 
 import com.hedvig.paymentservice.configuration.MerchantAccounts
-import com.hedvig.paymentservice.query.member.entities.Member
-import com.hedvig.paymentservice.query.member.entities.MemberRepository
+import com.hedvig.paymentservice.query.adyenAccount.AdyenAccount
+import com.hedvig.paymentservice.query.adyenAccount.AdyenAccountRepository
 import com.hedvig.paymentservice.serviceIntergration.memberService.MemberService
 import com.hedvig.paymentservice.serviceIntergration.productPricing.ProductPricingService
 import com.hedvig.paymentservice.serviceIntergration.productPricing.dto.ContractMarketInfo
@@ -19,117 +19,116 @@ import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.springframework.test.context.junit4.SpringRunner
-import java.util.Optional
+import java.util.UUID
 import javax.money.Monetary
 
 @RunWith(SpringRunner::class)
 class AdyenMerchantPickerTest() {
-  @MockkBean
-  lateinit var memberService: MemberService
+    @MockkBean
+    lateinit var memberService: MemberService
 
-  @MockkBean
-  lateinit var underwriterService: UnderwriterService
+    @MockkBean
+    lateinit var underwriterService: UnderwriterService
 
-  @MockkBean
-  lateinit var productPricingService: ProductPricingService
+    @MockkBean
+    lateinit var productPricingService: ProductPricingService
 
-  @MockkBean
-  lateinit var memberRepository: MemberRepository
+    @MockkBean
+    lateinit var merchantAccounts: MerchantAccounts
 
-  @MockkBean
-  lateinit var merchantAccounts: MerchantAccounts
+    @MockkBean
+    lateinit var adyenAccountRepository: AdyenAccountRepository
 
-  lateinit var adyenMerchantPicker: AdyenMerchantPicker
+    lateinit var adyenMerchantPicker: AdyenMerchantPicker
 
-  @Before
-  fun setup() {
-    adyenMerchantPicker = AdyenMerchantPicker(
-      memberService = memberService,
-      underwriterService = underwriterService,
-      productPricingService = productPricingService,
-      memberRepository = memberRepository,
-      merchantAccounts = makeMerchantAccount()
-    )
+    @Before
+    fun setup() {
+        adyenMerchantPicker = AdyenMerchantPicker(
+            memberService = memberService,
+            underwriterService = underwriterService,
+            productPricingService = productPricingService,
+            adyenAccountRepository = adyenAccountRepository,
+            merchantAccounts = makeMerchantAccount()
+        )
 
-    every { memberRepository.findById(any()) } returns Optional.empty()
-  }
+        every { adyenAccountRepository.findById(any()) } returns Optional.empty()
+    }
 
-  @Test
-  fun `Given payment-service, contracts and quotes failed to extract the market, get the market info from picked Locale`() {
-    every { memberService.getPickedLocale(any()) } returns "en_DK"
-    val x = adyenMerchantPicker.getAdyenMerchantInfo("1234")
+    @Test
+    fun `Given payment-service, contracts and quotes failed to extract the market, get the market info from picked Locale`() {
+        every { memberService.getPickedLocale(any()) } returns "en_DK"
+        val x = adyenMerchantPicker.getAdyenMerchantInfo("1234")
 
-    assertThat(x).isEqualTo(
-      AdyenMerchantInfo(
-        account = "HedvigTestDenmark",
-        countryCode = CountryCode.DK,
-        currencyCode = CurrencyCode.DKK
-      )
-    )
-  }
+        assertThat(x).isEqualTo(
+            AdyenMerchantInfo(
+                account = "HedvigTestDenmark",
+                countryCode = CountryCode.DK,
+                currencyCode = CurrencyCode.DKK
+            )
+        )
+    }
 
-  @Test
-  fun `Given only payment-service can extract the market, get the market info from member entity`() {
-    val member = Member()
-    member.adyenMerchantAccount = "HedvigTestDenmark"
+    @Test
+    fun `Given only payment-service can extract the market, get the market info from member entity`() {
+        val account = AdyenAccount("1234", "HedvigTestDenmark")
 
-    every { memberRepository.findById(any()) } returns Optional.of(member)
+        every { adyenAccountRepository.findById(any()) } returns Optional.of(account)
 
-    val x = adyenMerchantPicker.getAdyenMerchantInfo("1234")
+        val x = adyenMerchantPicker.getAdyenMerchantInfo("1234")
 
-    assertThat(x).isEqualTo(
-      AdyenMerchantInfo(
-        account = "HedvigTestDenmark",
-        countryCode = CountryCode.DK,
-        currencyCode = CurrencyCode.DKK
-      )
-    )
-  }
+        assertThat(x).isEqualTo(
+            AdyenMerchantInfo(
+                account = "HedvigTestDenmark",
+                countryCode = CountryCode.DK,
+                currencyCode = CurrencyCode.DKK
+            )
+        )
+    }
 
-  @Test
-  fun `Given only contact&aggrement can extract the market, get the market info from ContractMarketInfo`() {
-    every { productPricingService.getContractMarketInfo(any()) } returns ContractMarketInfo(
-      Market.DENMARK,
-      preferredCurrency = Monetary.getCurrency("DKK")
-    )
+    @Test
+    fun `Given only contact&aggrement can extract the market, get the market info from ContractMarketInfo`() {
+        every { productPricingService.getContractMarketInfo(any()) } returns ContractMarketInfo(
+            Market.DENMARK,
+            preferredCurrency = Monetary.getCurrency("DKK")
+        )
 
-    val x = adyenMerchantPicker.getAdyenMerchantInfo("1234")
+        val x = adyenMerchantPicker.getAdyenMerchantInfo("1234")
 
-    assertThat(x).isEqualTo(
-      AdyenMerchantInfo(
-        account = "HedvigTestDenmark",
-        countryCode = CountryCode.DK,
-        currencyCode = CurrencyCode.DKK
-      )
-    )
-  }
+        assertThat(x).isEqualTo(
+            AdyenMerchantInfo(
+                account = "HedvigTestDenmark",
+                countryCode = CountryCode.DK,
+                currencyCode = CurrencyCode.DKK
+            )
+        )
+    }
 
-  @Test
-  fun `Given only quotes can extract the market, get the market info from QuoteMarketInfo`() {
-    every { underwriterService.getMarketFromQuote(any()) } returns QuoteMarketInfo(
-      "DENMARK"
-    )
+    @Test
+    fun `Given only quotes can extract the market, get the market info from QuoteMarketInfo`() {
+        every { underwriterService.getMarketFromQuote(any()) } returns QuoteMarketInfo(
+            "DENMARK"
+        )
 
-    val x = adyenMerchantPicker.getAdyenMerchantInfo("1234")
+        val x = adyenMerchantPicker.getAdyenMerchantInfo("1234")
 
-    assertThat(x).isEqualTo(
-      AdyenMerchantInfo(
-        account = "HedvigTestDenmark",
-        countryCode = CountryCode.DK,
-        currencyCode = CurrencyCode.DKK
-      )
-    )
-  }
+        assertThat(x).isEqualTo(
+            AdyenMerchantInfo(
+                account = "HedvigTestDenmark",
+                countryCode = CountryCode.DK,
+                currencyCode = CurrencyCode.DKK
+            )
+        )
+    }
 
-  private fun makeMerchantAccount(): MerchantAccounts {
-    val merchantAccounts = MerchantAccounts()
-    merchantAccounts.merchantAccounts = mapOf<String, String>(
-      Pair("NORWAY", "HedvigTestNorway"),
-      Pair("DENMARK", "HedvigTestDenmark"),
-      Pair("SWEDEN", "HedvigTestSweden")
-    )
+    private fun makeMerchantAccount(): MerchantAccounts {
+        val merchantAccounts = MerchantAccounts()
+        merchantAccounts.merchantAccounts = mapOf<String, String>(
+            Pair("NORWAY", "HedvigTestNorway"),
+            Pair("DENMARK", "HedvigTestDenmark"),
+            Pair("SWEDEN", "HedvigTestSweden")
+        )
 
-    return merchantAccounts
-  }
+        return merchantAccounts
+    }
 }
